// src/app/pages/app-rrhh/documentos/models/vacacion-doc-new-format-dto.ts
export class VacacionDocNewFormatDto {
    employee: string;
    barcode: string;
    employeeId: number;
    job: string;
    department: string;
    fechaIngreso: string;
    cantidadDiasVacacion: number;
    fechaInicio: string;
    fechaFin: string;
    observaciones: string;
    actividadesPendientes: string;
    cubreVacaciones: string;
    fechaReintegro: string;
    jefeInmediato: string;
    tipoVacacionId: number;
    tipoVacacionName: string;
    cantidadDiasVacacionEnLetras: string;
}